fix : 검증 조건 추가

+++ src/main/kotlin/com/tovelop/maphant/controller/CommentController.kt
@@ -5,7 +5,6 @@ import com.tovelop.maphant.service.CommentService
 import com.tovelop.maphant.type.response.Response
 import com.tovelop.maphant.type.response.ResponseUnit
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.data.annotation.Id
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
 
@@ -66,23 +65,20 @@ class CommentController(@Autowired val commentService: CommentService) {
         if (commentService.getCommentById(commentDTO.id)!!.is_anonymous != commentDTO.is_anonymous) {
             return ResponseEntity.badRequest().body(Response.error("댓글 익명 여부는 수정할 수 없습니다."))
         }
-        if (commentService.getCommentById(commentDTO.id)!!.like_cnt != commentDTO.like_cnt) {
-            return ResponseEntity.badRequest().body(Response.error("댓글 좋아요 수는 수정할 수 없습니다."))
-        }
-        if (commentService.getCommentById(commentDTO.id)!!.id != commentDTO.id) {
-            return ResponseEntity.badRequest().body(Response.error("댓글 번호는 수정할 수 없습니다."))
-        }
         commentService.updateComment(commentDTO)
         return ResponseEntity.ok().body(Response.stateOnly(true))
     }
 
     @PostMapping("/insert-like")
     fun insertCommentLike(@RequestBody commentRequest: CommentRequest): ResponseEntity<ResponseUnit> {
+        if (commentService.findCommentLike(commentRequest.userId, commentRequest.commentId) != null) {
+            return ResponseEntity.badRequest().body(Response.error("이미 좋아요를 누른 댓글입니다."))
+        }
         commentService.insertCommentLike(commentRequest.userId, commentRequest.commentId)
         return ResponseEntity.ok().body(Response.stateOnly(true))
     }
 
-    @PostMapping("/find-like") // TODO : 수정
+    @PostMapping("/find-like")
     fun findCommentLike(@RequestBody commentRequest: CommentRequest): ResponseEntity<Response<Int>> {
         commentService.findCommentLike(commentRequest.userId, commentRequest.commentId)
         return ResponseEntity.ok().body(Response.success(commentRequest.commentId))
@@ -90,13 +86,18 @@ class CommentController(@Autowired val commentService: CommentService) {
 
     @GetMapping("/cnt-like/{commentId}")
     fun cntCommentLike(@PathVariable commentId: Int): ResponseEntity<ResponseUnit> {
+        if (commentService.getCommentById(commentId) == null) {
+            return ResponseEntity.badRequest().body(Response.error("존재하지 않는 댓글입니다."))
+        }
         commentService.cntCommentLike(commentId)
-        println(commentService.cntCommentLike(commentId))
         return ResponseEntity.ok().body(Response.stateOnly(true))
     }
 
     @PostMapping("/delete-like")
     fun deleteCommentLike(@RequestBody commentRequest: CommentRequest): ResponseEntity<ResponseUnit> {
+        if (commentService.findCommentLike(commentRequest.userId, commentRequest.commentId) == null) {
+            return ResponseEntity.badRequest().body(Response.error("좋아요를 누르지 않은 댓글입니다."))
+        }
         commentService.deleteCommentLike(commentRequest.userId, commentRequest.commentId)
         return ResponseEntity.ok().body(Response.stateOnly(true))
     }
@@ -106,11 +107,14 @@ class CommentController(@Autowired val commentService: CommentService) {
         if (commentService.findCommentReport(commentRequest.userId, commentRequest.commentId) != null) {
             return ResponseEntity.badRequest().body(Response.error("이미 신고한 댓글입니다."))
         }
+        if (commentRequest.reportId == null) {
+            return ResponseEntity.badRequest().body(Response.error("신고 사유를 선택해주세요."))
+        }
         commentService.insertCommentReport(commentRequest.userId, commentRequest.commentId, commentRequest.reportId!!)
         return ResponseEntity.ok().body(Response.stateOnly(true))
     }
 
-    @PostMapping("/find-report") // TODO : 수정
+    @PostMapping("/find-report")
     fun findCommentReport(@RequestBody commentRequest: CommentRequest): ResponseEntity<ResponseUnit> {
         commentService.findCommentReport(commentRequest.userId, commentRequest.commentId)
         return ResponseEntity.ok().body(Response.stateOnly(true))
